pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ato-mcp"
version = "0.1.0"
description = "Local MCP server for the Australian Taxation Office legal database"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Jordan" }]
keywords = ["mcp", "ato", "tax", "australia", "legal", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.2",
    "fastmcp>=2.0",
    "typer>=0.12",
    "httpx[http2]>=0.27",
    "selectolax>=0.3.21",
    "markdownify>=0.11",
    "sqlite-vec>=0.1.6",
    "onnxruntime>=1.20",
    "tokenizers>=0.20",
    "orjson>=3.10",
    "zstandard>=0.22",
    "platformdirs>=4",
    "pydantic>=2",
    "numpy>=1.26",
    "pyyaml>=6",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]
# Signature verification is opt-in. Install via `pip install 'ato-mcp[verify]'`
# to enable manifest checking; the update path logs a warning and proceeds
# without it when the extra is missing.
verify = [
    "minisign>=0.12",
]

[project.scripts]
ato-mcp = "ato_mcp.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ato_mcp"]

[tool.hatch.build.targets.wheel.force-include]
"data/doc_type_map.yaml" = "ato_mcp/_data/doc_type_map.yaml"
"data/eval_queries.yaml" = "ato_mcp/_data/eval_queries.yaml"

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
